fix: change components&class namings, little change for howMuchSpent input validation, fix code style feat: add helpers folder with convertDate function
