+++ src/components/Cost/index.jsx
@@ -1,3 +1,4 @@
+import { convertDate } from "../../helpers";
 import editIcon from "../../img/edit-icon.svg";
 import deleteIcon from "../../img/delete-icon.svg";
 import "./style.scss";
@@ -6,8 +7,7 @@ const Cost = ({
   cost,
   index,
   startEdit,
-  deleteCost,
-  convertDate
+  deleteCost
 }) => {
   const { whereSpent, whenSpent, howMuchSpent, id } = cost;
 
@@ -17,48 +17,33 @@ const Cost = ({
   return (
     <div className="cost">
       <div className="cost-info">
-        <input 
-          type="text"
-          className="cost-info__text cost-where-spent"
-          value={`${index + 1}) ` + whereSpent}
-          readOnly
-        />
-        <input 
-          type="text"
-          className="cost-info__text cost-when-spent"
-          value={dateWhenSpent}
-          readOnly
-        />
-        <input 
-          type="text"
-          className="cost-info__text cost-how-much-spent"
-          value={`${howMuchSpent} р.`}
-          readOnly
-        />
+        <p className="cost-info__text cost-where-spent">{`${index + 1}) ` + whereSpent}</p>
+        <p className="cost-info__text cost-when-spent">{dateWhenSpent}</p>
+        <p className="cost-info__text cost-how-much-spent">{`${howMuchSpent} р.`}</p>
       </div>
       <div className="cost-actions">
         <button 
           type="button"
           className="cost-actions__button"
           onClick={() => startEdit(cost)}
-          >
-            <img 
-              src={editIcon}
-              alt="" 
-              className="cost-actions__img"
-            />
-          </button>
-          <button 
+        >
+          <img 
+            src={editIcon}
+            alt="" 
+            className="cost-actions__img"
+          />
+        </button>
+        <button 
           type="button"
           className="cost-actions__button"
           onClick={() => deleteCost(id)}
-          >
-            <img 
-              src={deleteIcon}
-              alt="" 
-              className="cost-actions__img"
-            />
-          </button>
+        >
+          <img 
+            src={deleteIcon}
+            alt="" 
+            className="cost-actions__img"
+          />
+        </button>
       </div>
     </div>
   )
