+++ src/components/Cost/style.scss
@@ -13,7 +13,8 @@
       border: 0;
       background: #C5E9FF 50%;
       word-wrap: break-word;
-      margin-right: 2vw;
+      margin: 0 2vw 0 0;
+      padding: 0;
     }
   }
