+++ src/components/MainPage/style.scss
@@ -1,5 +1,5 @@
-.container {
-  width: 100%;
+.main-page {
+  width: 89%;
   display: flex;
   flex-direction: column;
