--- src/components/TotalCost/index.jsx
@@ -1,9 +1,0 @@
-import "./style.scss";
-
-const TotalCost = ({ totalSum }) => {
-  return (
-    <p className="total-cost">{`Итого: ${totalSum} р.`}</p>
-  )
-}
-
-export default TotalCost;
