--- src/components/TotalCost/style.scss
@@ -1,4 +1,0 @@
-.total-cost {
-  align-self: flex-end;
-  font-size: 3rem;
-}
