+++ src/components/TotalPrice/style.scss
@@ -1,10 +1,10 @@
-.total-cost {
+.total-price {
   align-self: flex-end;
   font-size: 3rem;
 }
 
 @media screen and (max-width: 1400px) {
-  .total-cost {
+  .total-price {
     font-size: 2.5rem;
   }
 }
